module: updte

+++ include/DrawWrapper.hpp
@@ -4,7 +4,7 @@
 
 #include <vector>
 #include "Widget.hpp"
-#include "Types.hpp"
+
 namespace GUI
 {
     extern const bool isRelative;
@@ -16,8 +16,8 @@ namespace GUI
     void draw_triangle(DrawPoint A, DrawPoint B, DrawPoint C);
     void draw_text(int row, int column, mstd::wstring, Color color);
     void draw_textShadow(int row, int column, mstd::wstring, Color color);
-    void draw_translate(int x, int y);
-    void _draw_translate(int x, int y);
+    void draw_translate(float x, float y);
+    void _draw_translate(float x, float y);
 }
 
 #endif
